Removed the TouchdownDetectorBasedFootSwitch and updated the QuadrupedFootSwitchFactory.
==========================================

There is still a hardcoded parameter here!! Please move to a proper
parameter file

contactThresholdForce is set in setupTouchdownBasedFootSwitches

==========================================

The TouchdownDetectorBasedFootSwitch was a way to fuse different
touchdown detection methods but it ended up being too naive to be
useful. ComputedForceBasedFootSwitch has been effective when no contact
sensors are available. Physical Contact sensors are recommended.